requirements.txt
openai
scikit-learn
matplotlib
seaborn
python-dotenv